+++ mr_robot/Cogs/translate.py
@@ -1,11 +1,12 @@
 import logging
 
 import disnake
-import googletrans as gt
+import googletrans as gt  # TODO: replace googletrans with its alternative & update httpx to latest version
 from disnake.ext import commands
 from googletrans import Translator
-
 from utils import Embeds, delete_button
+
+from mr_robot.bot import MrRobot
 
 
 async def autocomp_langs(inter: disnake.CommandInteraction, user_input: str):
@@ -19,7 +20,7 @@ logger = logging.getLogger(__name__)
 
 
 class Translate(commands.Cog):
-    def __init__(self, client):
+    def __init__(self, client: MrRobot):
         self.bot = client
         logger.info("Translate Cog loaded")
 
@@ -50,5 +51,5 @@ class Translate(commands.Cog):
         )
 
 
-def setup(client: commands.Bot):
+def setup(client: MrRobot):
     client.add_cog(Translate(client))
