+++ mr_robot/bot.py
@@ -4,35 +4,12 @@ import os
 import time
 from typing import Dict, List
 
-import aiosqlite
-import disnake
 import httpx
 import mafic
 from aiocache import cached
 from disnake.ext import commands
-from dotenv import load_dotenv
-
 from git_api import Git
-from utils import proxy_generator
 
-proxy_mode = False
-PROXY = None
-
-file_handler = logging.FileHandler("mr-robot.log", mode="w")
-console_handler = logging.StreamHandler()
-
-file_handler.setLevel(logging.INFO)
-console_handler.setLevel(logging.INFO)
-logging.basicConfig(
-    level=logging.NOTSET,
-    format="%(levelname)s - %(name)s - %(filename)s - %(module)s - %(funcName)s - %(message)s",
-    handlers=[console_handler, file_handler],
-)
-
-logging.getLogger("httpx").setLevel(logging.WARNING)
-logging.getLogger("httpcore").setLevel(logging.WARNING)
-logging.getLogger("disnake").setLevel(logging.INFO)
-logging.getLogger("aiosqlite").setLevel(logging.INFO)
 logger = logging.getLogger(__name__)
 
 
@@ -92,7 +69,7 @@ class MrRobot(commands.AutoShardedInteractionBot):
                 exp_pow += 1
                 logger.warning(f"Trying to reload player after {2**exp_pow} seconds")
 
-    def load_extensions(self):
+    def load_bot_extensions(self):
         """Loads extensions"""
         for file in os.listdir("Cogs"):
             if file.endswith(".py"):
@@ -100,37 +77,3 @@ class MrRobot(commands.AutoShardedInteractionBot):
                     self.load_extension(f"Cogs.{str(file[:-3])}")
                 except Exception as error:
                     raise error
-
-
-load_dotenv()
-
-
-async def main():
-    global PROXY
-    db_name = "mr-robot.db"
-    async with httpx.AsyncClient(timeout=httpx.Timeout(None)) as session:
-        client = MrRobot(
-            proxy=PROXY,
-            intents=disnake.Intents.all(),
-            session=session,
-            db=await aiosqlite.connect(db_name),
-            db_name=db_name,
-        )
-        if client.git:
-            logger.info("Pulling DB")
-            await client.git.pull(db_name)
-        client.load_extensions()
-        try:
-            await client.start(os.getenv("BOT_TOKEN"))  # type: ignore
-        except (disnake.errors.LoginFailure, disnake.errors.HTTPException):
-            logger.warning(
-                "Unable to connect to Discord falling back to proxy mode", exc_info=True
-            )
-            proxy_mode = True
-            PROXY = proxy_generator() if proxy_mode else None
-        finally:
-            await client.close()
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
